Move catch_errors_wrapper to convert_frame
With this change, we now have the invariant that eval_frame only
contains "hot" functions that are called at runtime, as opposed to
cold functions which are only called at compile time.


[ghstack-poisoned]
